Use Float32Array types in .xkt v3, not Float64Array

+++ src/lib/math.js
@@ -1,9 +1,9 @@
 
 // Some temporary vars to help avoid garbage collection
 
-const tempMat1 = new Float64Array(16);
-const tempMat2 = new Float64Array(16);
-const tempVec4 = new Float64Array(4);
+const tempMat1 = new Float32Array(16);
+const tempMat2 = new Float32Array(16);
+const tempVec4 = new Float32Array(4);
 
 /**
  * @private
@@ -35,7 +35,7 @@ const math = {
      * @returns {Number[]}
      */
     vec2(values) {
-        return new Float64Array(values || 2);
+        return new Float32Array(values || 2);
     },
 
     /**
@@ -46,7 +46,7 @@ const math = {
      * @returns {Number[]}
      */
     vec3(values) {
-        return new Float64Array(values || 3);
+        return new Float32Array(values || 3);
     },
 
     /**
@@ -57,7 +57,7 @@ const math = {
      * @returns {Number[]}
      */
     vec4(values) {
-        return new Float64Array(values || 4);
+        return new Float32Array(values || 4);
     },
 
     /**
@@ -68,7 +68,7 @@ const math = {
      * @returns {Number[]}
      */
     mat3(values) {
-        return new Float64Array(values || 9);
+        return new Float32Array(values || 9);
     },
 
     /**
@@ -79,7 +79,7 @@ const math = {
      * @static
      * @returns {Number[]}
      */
-    mat3ToMat4(mat3, mat4 = new Float64Array(16)) {
+    mat3ToMat4(mat3, mat4 = new Float32Array(16)) {
         mat4[0] = mat3[0];
         mat4[1] = mat3[1];
         mat4[2] = mat3[2];
@@ -107,7 +107,7 @@ const math = {
      * @returns {Number[]}
      */
     mat4(values) {
-        return new Float64Array(values || 16);
+        return new Float32Array(values || 16);
     },
 
     /**
@@ -119,7 +119,7 @@ const math = {
      * @returns {Number[]}
      */
     mat4ToMat3(mat4, mat3) { // TODO
-        //return new Float64Array(values || 9);
+        //return new Float32Array(values || 9);
     },
 
     /**
@@ -682,7 +682,7 @@ const math = {
     },
 
     distVec3: ((() => {
-        const vec = new Float64Array(3);
+        const vec = new Float32Array(3);
         return (v, w) => math.lenVec3(math.subVec3(v, w, vec));
     }))(),
 
@@ -698,7 +698,7 @@ const math = {
     },
 
     distVec2: ((() => {
-        const vec = new Float64Array(2);
+        const vec = new Float32Array(2);
         return (v, w) => math.lenVec2(math.subVec2(v, w, vec));
     }))(),
 
@@ -768,7 +768,7 @@ const math = {
      */
     vec3FromMat4Scale: ((() => {
 
-        const tempVec3 = new Float64Array(3);
+        const tempVec3 = new Float32Array(3);
 
         return (m, dest) => {
 
@@ -828,7 +828,7 @@ const math = {
      * @returns {*[]}
      */
     xyzObjectToArray(xyz, arry) {
-        arry = arry || new Float64Array(3);
+        arry = arry || new Float32Array(3);
         arry[0] = xyz.x;
         arry[1] = xyz.y;
         arry[2] = xyz.z;
@@ -895,7 +895,7 @@ const math = {
      * @static
      */
     diagonalMat4v(v) {
-        return new Float64Array([
+        return new Float32Array([
             v[0], 0.0, 0.0, 0.0,
             0.0, v[1], 0.0, 0.0,
             0.0, 0.0, v[2], 0.0,
@@ -926,7 +926,7 @@ const math = {
      * @method identityMat4
      * @static
      */
-    identityMat4(mat = new Float64Array(16)) {
+    identityMat4(mat = new Float32Array(16)) {
         mat[0] = 1.0;
         mat[1] = 0.0;
         mat[2] = 0.0;
@@ -955,7 +955,7 @@ const math = {
      * @method identityMat3
      * @static
      */
-    identityMat3(mat = new Float64Array(9)) {
+    identityMat3(mat = new Float32Array(9)) {
         mat[0] = 1.0;
         mat[1] = 0.0;
         mat[2] = 0.0;
@@ -1235,7 +1235,7 @@ const math = {
      */
     mulMat3(a, b, dest) {
         if (!dest) {
-            dest = new Float64Array(9);
+            dest = new Float32Array(9);
         }
 
         const a11 = a[0];
@@ -1538,7 +1538,7 @@ const math = {
      * @static
      */
     translationMat4c: ((() => {
-        const xyz = new Float64Array(3);
+        const xyz = new Float32Array(3);
         return (x, y, z, dest) => {
             xyz[0] = x;
             xyz[1] = y;
@@ -1716,7 +1716,7 @@ const math = {
      * @static
      */
     scalingMat4c: ((() => {
-        const xyz = new Float64Array(3);
+        const xyz = new Float32Array(3);
         return (x, y, z, dest) => {
             xyz[0] = x;
             xyz[1] = y;
@@ -1948,8 +1948,8 @@ const math = {
 
     decomposeMat4: (() => {
 
-        const vec = new Float64Array(3);
-        const matrix = new Float64Array(16);
+        const vec = new Float32Array(3);
+        const matrix = new Float32Array(16);
 
         return function decompose(mat, position, quaternion, scale) {
 
@@ -2598,9 +2598,9 @@ const math = {
      * @static
      */
     unprojectVec3: ((() => {
-        const mat = new Float64Array(16);
-        const mat2 = new Float64Array(16);
-        const mat3 = new Float64Array(16);
+        const mat = new Float32Array(16);
+        const mat2 = new Float32Array(16);
+        const mat3 = new Float32Array(16);
         return function (p, viewMat, projMat, q) {
             return this.transformVec3(this.mulMat4(this.inverseMat4(viewMat, mat), this.inverseMat4(projMat, mat2), mat3), p, q)
         };
@@ -2832,7 +2832,7 @@ const math = {
     },
 
     quaternionToEuler: ((() => {
-        const mat = new Float64Array(16);
+        const mat = new Float32Array(16);
         return (q, order, dest) => {
             dest = dest || math.vec3();
             math.quaternionToRotationMat4(q, mat);
@@ -3018,7 +3018,7 @@ const math = {
      * @private
      */
     AABB3(values) {
-        return new Float64Array(values || 6);
+        return new Float32Array(values || 6);
     },
 
     /**
@@ -3027,7 +3027,7 @@ const math = {
      * @private
      */
     AABB2(values) {
-        return new Float64Array(values || 4);
+        return new Float32Array(values || 4);
     },
 
     /**
@@ -3036,7 +3036,7 @@ const math = {
      * @private
      */
     OBB3(values) {
-        return new Float64Array(values || 32);
+        return new Float32Array(values || 32);
     },
 
     /**
@@ -3045,12 +3045,12 @@ const math = {
      * @private
      */
     OBB2(values) {
-        return new Float64Array(values || 16);
+        return new Float32Array(values || 16);
     },
 
     /** Returns a new 3D bounding sphere */
     Sphere3(x, y, z, r) {
-        return new Float64Array([x, y, z, r]);
+        return new Float32Array([x, y, z, r]);
     },
 
     /**
@@ -3118,9 +3118,9 @@ const math = {
      */
     getAABB3Diag: ((() => {
 
-        const min = new Float64Array(3);
-        const max = new Float64Array(3);
-        const tempVec3 = new Float64Array(3);
+        const min = new Float32Array(3);
+        const max = new Float32Array(3);
+        const tempVec3 = new Float32Array(3);
 
         return aabb => {
 
@@ -3145,9 +3145,9 @@ const math = {
      */
     getAABB3DiagPoint: ((() => {
 
-        const min = new Float64Array(3);
-        const max = new Float64Array(3);
-        const tempVec3 = new Float64Array(3);
+        const min = new Float32Array(3);
+        const max = new Float32Array(3);
+        const tempVec3 = new Float32Array(3);
 
         return (aabb, p) => {
 
@@ -3279,7 +3279,7 @@ const math = {
      */
     positions3ToAABB3: ((() => {
 
-        const p = new Float64Array(3);
+        const p = new Float32Array(3);
 
         return (positions, aabb, positionsDecodeMatrix) => {
             aabb = aabb || math.AABB3();
@@ -3474,7 +3474,7 @@ const math = {
      */
     points3ToSphere3: ((() => {
 
-        const tempVec3 = new Float64Array(3);
+        const tempVec3 = new Float32Array(3);
 
         return (points, sphere) => {
 
@@ -3522,8 +3522,8 @@ const math = {
      */
     positions3ToSphere3: ((() => {
 
-        const tempVec3a = new Float64Array(3);
-        const tempVec3b = new Float64Array(3);
+        const tempVec3a = new Float32Array(3);
+        const tempVec3b = new Float32Array(3);
 
         return (positions, sphere) => {
 
@@ -3577,8 +3577,8 @@ const math = {
      */
     OBB3ToSphere3: ((() => {
 
-        const point = new Float64Array(3);
-        const tempVec3 = new Float64Array(3);
+        const point = new Float32Array(3);
+        const tempVec3 = new Float32Array(3);
 
         return (points, sphere) => {
